+++ test/dataset_tests.jl
@@ -19,6 +19,18 @@ using Base.Test, StaticArrays
     end
   end
 
+  @testset "Conversions" begin
+    m = Matrix(data)
+    @test Dataset(m) == data
+    @test reinterpret(Dataset, reinterpret(Matrix, data)) == data
+    @test transpose(m) == reinterpret(Matrix, data)
+
+    m = rand(1000, 4)
+    @test reinterpret(Matrix, reinterpret(Dataset, m)) == m
+    @test Dataset(m) !== Dataset(transpose(m))
+    @test Dataset(m) == reinterpret(Dataset, transpose(m))
+  end
+
   @testset "IO" begin
     @test !isfile("test.txt")
     write_dataset("test.txt", data)
@@ -32,6 +44,19 @@ using Base.Test, StaticArrays
     @test dimension(data2) == 2
 
     rm("test.txt")
+
+    write_dataset("test.txt", data, ',')
+    @test isfile("test.txt")
+
+    data3 = read_dataset("test.txt", Dataset{3, Float64}, ',')
+    @test dimension(data3) == 3
+    @test data3 == data
+
+    data2 = read_dataset("test.txt", Dataset{2, Float64})
+    @test dimension(data2) == 2
+
+    rm("test.txt")
+
     @test !isfile("test.txt") # make extra sure!
   end
 end