+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.3.2
+* Orders of magnitude speed-up in conversions between `Matrix` and `Dataset`,
+  because now methods use `transpose` internally and only `reinterpret`.
+* Bugfix of `eltype` of `Reconstruction`.
+
 # v0.3.1
 * Added `jacobian` function
 * Removed `EomVector` nonsense.